weather_app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
requests==2.28.1
python-dotenv==0.19.0
mysql-connector-python==8.0.28
